Slight reworking of the homepage.
I put the top image and quote in a hero unit and made it so the quote isn't
using an h* tag. I also reworked the blurbs so that side-by-side blurbs
line up. I put the final info-note at the end in a well to give it some
contrast and took out the center align.